files (3)/requirements.txt
pdf2image>=1.16.0
Pillow>=9.0.0
pywin32>=305
